feature: subscription list

+++ receivedMessage/receivedPostback.js
@@ -1,5 +1,6 @@
 var sendTextMessage = require('../sendMessage/sendTextMessage');
 var qrFavorite = require('../sendMessage/qrFavorite');
+var subscriptionList = require('../sendMessage/subscriptionList');
 var schedule = require('node-schedule');
 /*
 * Postback Event
@@ -24,21 +25,46 @@ function receivedPostback(event, db, callback) {
   // let them know it was successful
   switch (payload) {
     case "How_to_use":
-    var intro = '嗨! =D 我是TenkiBot☀☁☔\n我可以幫您~\n查詢今明預報:\n?縣市名稱\n\n查詢一週預報:\n!縣市名稱\n\n訂閱(自動定時推播天氣資訊):\n#縣市名稱 時間\n(ex: #台北 700 \nor #新竹 2330)\n\n*想移除所有訂閱請輸入"@"即可, 或是直接按選單中的"取消訂閱"\n\n將城市加入最愛(方便查詢):\n+縣市名稱\n\n移除最愛:\n-縣市名稱\n(也可以使用選單中的"移除最愛"喔)';
+    var intro = '嗨! =D 我是TenkiBot☀☁☔\n依照以下格式輸入訊息我可以幫您\n查詢今明天氣預報:\n?縣市名稱\n\n查詢一週天氣預報:\n!縣市名稱\n\n訂閱(定時自動推播天氣資訊):\n#縣市名稱 幾點幾分\n(ex: #台北 700 \nor #新竹 2330)\n\n移除所有訂閱請輸入"x"或按選單中的"取消訂閱"\n\n將城市加入最愛(方便查詢):\n+縣市名稱\n\n移除最愛:\n-縣市名稱\n(或按選單中的"移除最愛")';
     sendTextMessage(senderID, intro);
     break;
 
     case "Subscription":
     // var intro = "請設定自動推播時間(ex:7:00, 22:10):"
-    var intro = '請輸入: #縣市名稱 時間\n(例如@台北 700 訂閱每天7點傳送台北天氣\nor #新竹 2330 每天23點30分傳送新竹天氣)\n\n想移除所有訂閱請輸入"@"即可, 或是直接按選單中的"取消訂閱"\n';
-    sendTextMessage(senderID, intro);
+    var intro = 'p.s.訂閱方法說明: 輸入 #縣市名稱 幾點幾分\n(例如#台北 700 訂閱每天7點傳送台北天氣\nor #新竹 2330 每天23點30分傳送新竹天氣)';
+    db.collection('subscription').find({'senderID': senderID}).toArray((err, res)=>{
+      if(err) return console.log('\subscription find db error: %s\n'+err);
+      if (res.length==0) {
+        sendTextMessage(senderID, '您尚未開始訂閱\n'+intro);
+        return;
+      }
+      var subCitys = [{
+        "title": '您目前已訂閱:',
+        "subtitle": '',
+        "buttons": []
+      }];
+      for (var i = 0; i < res.length; i++) {
+        subCitys[0]["buttons"].push({
+          "type":"postback",
+          "title": i+1+'. '+res[i].cityName+' '+(res[i].hour<10? '0':'')+res[i].hour+':'+(res[i].minute<10? '0':'')+res[i].minute,
+          "payload":"remove_single_sub"
+        })
+      }
+      subCitys[0]["buttons"].push({
+        "type":"postback",
+        "title":"取消全部訂閱",
+        "payload":"remove_subscription"
+      });
+      subscriptionList(senderID, subCitys);
+    });
+    callback();
     break;
 
     case "Favorite":
     db.collection('favorite').find({'senderID': senderID}).toArray((err, res)=>{
       if(err) return console.log('\nfavorite find db error: %s\n'+err);
       if (res.length==0) {
-        sendTextMessage(senderID, '還沒有城市被加入喔\n(將常用的城市加入最愛以後可以更方便查詢!)');
+        sendTextMessage(senderID, '還沒有城市被加入喔\n(使用"+縣市名稱"將常用的城市加入最愛可以更方便查詢!)');
         return;
       }
       var faCitys = [];
@@ -59,7 +85,7 @@ function receivedPostback(event, db, callback) {
     db.collection('favorite').find({'senderID': senderID}).toArray((err, res)=>{
       if(err) return console.log('favorite find db error: %s\n'+err);
       if (res.length==0) {
-        sendTextMessage(senderID, '還沒有城市被加入喔(將常用的城市加入最愛可以更方便查詢!)');
+        sendTextMessage(senderID, '還沒有城市被加入喔\n(使用"+縣市名稱"將常用的城市加入最愛可以更方便查詢!)');
         return;
       }
       var faCitys = [];
@@ -77,17 +103,28 @@ function receivedPostback(event, db, callback) {
     break;
 
     case 'remove_subscription':
-    db.collection('subscription').deleteMany({senderID: senderID}, (err, res)=>{
-      if(err) return console.log('\nremove subscription error');
-      console.log('\nremoved subscription '+res);
-    });
     var remove_scheduleJob = schedule.scheduledJobs[senderID];
     try {
       remove_scheduleJob.cancel();
     } catch (e) {
-      console.log('schedule have canceled');
+      console.log('\n***schedule have canceled');
     }
-    sendTextMessage(senderID, '已幫您取消訂閱');
+    db.collection('subscription').find({senderID: senderID}).toArray((err, res)=>{
+      if(res.length>0){
+        db.collection('subscription').deleteMany({senderID: senderID}, (err, res)=>{
+          if(err) return console.log('\nremove subscription error');
+          console.log('\nremoved subscription '+res);
+          sendTextMessage(senderID, '已幫您取消訂閱');
+          return;
+        });
+      } else {
+        sendTextMessage(senderID, '您尚未開始訂閱');
+        return;
+      }
+    });
+
+    case 'remove_single_sub':
+    break;
 
     default:
     // sendTextMessage(senderID, payload);
